preparing dry up content
